Inject HUD "properly"
