+++ src/Glimpse.Agent.AspNet.Mvc/Providers/ScriptInjector.cs
@@ -1,4 +1,5 @@
 using System;
+using Glimpse.Web;
 using Microsoft.AspNet.Html.Abstractions;
 using Microsoft.AspNet.Mvc.Rendering;
 using Microsoft.AspNet.Razor.Runtime.TagHelpers;
@@ -9,29 +10,21 @@ namespace Glimpse.Agent.AspNet.Mvc
     public class ScriptInjector : TagHelper
     {
         private readonly Guid _requestId;
+        private readonly ScriptOptions _scriptOptions;
 
-        public ScriptInjector(IContextData<MessageContext> context)
+        public ScriptInjector(IContextData<MessageContext> context, IScriptOptionsProvider scriptOptionsProvider)
         {
             _requestId = context.Value.Id;
+            _scriptOptions = scriptOptionsProvider.BuildInstance();
         }
 
         public override int Order => int.MaxValue;
 
         public override void Process(TagHelperContext context, TagHelperOutput output)
         {
-            var js = @"
-var script = document.querySelector('script[data-request-id]'),
-link = document.createElement('a'),
-requestId = script.dataset.requestId;
-link.setAttribute('href', '/glimpse/client/index.html?id=' + requestId);
-link.setAttribute('target', '_blank');
-link.text = 'Open Glimpse';
-document.body.appendChild(link);";
-
             var tag = new TagBuilder("script");
-            tag.InnerHtml.SetContentEncoded(js);
-
             tag.Attributes.Add("data-request-id", _requestId.ToString());
+            tag.Attributes.Add("src", _scriptOptions.HudClientScriptUri);
 
             output.PostContent.SetContent(tag);
         }
